+++ src/main/java/ooad/service/impl/ModuleService.java
@@ -1,7 +1,8 @@
 package ooad.service.impl;
 
 import ooad.bean.*;
-import ooad.common.ModuleStatus;
+import ooad.common.Role;
+import ooad.common.exceptions.AuthorityException;
 import ooad.common.exceptions.NoSuchEntryException;
 import ooad.common.util.StringToSqlDate;
 import ooad.dao.ModuleDAO;
@@ -19,17 +20,26 @@ public class ModuleService implements IModuleService {
     ModuleDAO moduleDAO;
 
     @Override
-    public Module findModule(String module_name) {
-        return moduleDAO.findModule(module_name);
+    public List<Module> findModule(String module_name) {
+        try {
+            return moduleDAO.findModule(module_name);
+        } catch (NoSuchEntryException e) {
+            e.printStackTrace();
+            return null;
+        }
     }
 
     @Override
     public List<Module> getModules() {
-        return moduleDAO.getModules();
+//        return moduleDAO.getModules(x);
+        return null;
     }
 
     @Override
-    public Boolean newModule(String module_name, String module_content) {
+    public Boolean newModule(Role role,String module_name, String module_content)throws AuthorityException {
+        if( !role.equals(Role.Admin) ) {
+            throw new AuthorityException(role);
+        }
         Module newModule = new Module();
         newModule.setName(module_name);
         newModule.setDescription(module_content);
@@ -38,26 +48,38 @@ public class ModuleService implements IModuleService {
     }
 
     @Override
-    public Boolean deleteModule(int module_id) {
+    public Boolean deleteModule(Role role,int module_id)throws AuthorityException {
+        if( !role.equals(Role.Admin) ) {
+            throw new AuthorityException(role);
+        }
         boolean result = moduleDAO.deleteModule(module_id);
         return result;
     }
 
     @Override
-    public Boolean modifyModule(int module_id, String module_name, String module_content) {
+    public Boolean modifyModule(Role role,int module_id, String module_name, String module_content)throws AuthorityException {
+        if( !role.equals(Role.Admin) ) {
+            throw new AuthorityException(role);
+        }
         boolean result = moduleDAO.updateModule(module_id, module_name, module_content);
         return result;
     }
 
     @Override
-    public Boolean addAssignmentToModule(int module_id, int assignment_id) {
+    public Boolean addAssignmentToModule(Role role,int module_id, int assignment_id)throws AuthorityException {
+        if( !role.equals(Role.Admin) ) {
+            throw new AuthorityException(role);
+        }
         ModuleAssignment newEntry = new ModuleAssignment(module_id, assignment_id);
         boolean result = moduleDAO.addAssignmentToModule(newEntry);
         return result;
     }
 
     @Override
-    public Boolean deleteAssignmentFromModule(int module_id, int assignment_id) {
+    public Boolean deleteAssignmentFromModule(Role role,int module_id, int assignment_id)throws AuthorityException {
+        if( !role.equals(Role.Admin) ) {
+            throw new AuthorityException(role);
+        }
         boolean result = moduleDAO.deleteAssignmentToModule(module_id, assignment_id);
         return result;
     }
@@ -68,7 +90,10 @@ public class ModuleService implements IModuleService {
     }
 
     @Override
-    public Boolean publishModule(int module_id, int company_id, String start_time, String finish_time) {
+    public Boolean publishModule(Role role,int module_id, int company_id, String start_time, String finish_time) throws AuthorityException{
+        if( !role.equals(Role.Admin) ) {
+            throw new AuthorityException(role);
+        }
         Date start = StringToSqlDate.strToDate(start_time);
         Date finish = StringToSqlDate.strToDate(finish_time);
         ModuleProcess moduleProcess = new ModuleProcess(module_id, company_id, start, finish);
